My custom Post Processor for Haas ST-20Y
This post processor has a mode called ForceXyMilling. it uses the machine's X and Y axes like a milling machine would. it's very useful when making small parts, because the closer you get to the machine's rotation center, the faster the C axis needs to rotate in order to preserve the programmed feed rate